scraper base functionality

+++ core/include/scraper.h
@@ -7,6 +7,6 @@
 
 namespace Scraper
 {
-    void scanFloder(std::string path, int platform_id, const std::string& db_path);
+    void scanFloder(std::string path, long platform_id, const std::string& db_path);
     void findGamesInformation(const std::string& db_path);
 };
